--- src/extensions/document/document.ts
@@ -1,8 +1,0 @@
-import { Document as TiptapDocument } from "@tiptap/extension-document";
-
-export const Document = TiptapDocument.extend({
-  content: "block+",
-});
-
-export default Document;
-
