+++ src/hooks/useTemplateEditor.ts
@@ -2,22 +2,22 @@ import { Editor, EditorOptions, useEditor } from '@tiptap/react';
 import StarterKit from '@tiptap/starter-kit';
 import TaskList from '@tiptap/extension-task-list';
 import TaskItem from '@tiptap/extension-task-item';
-import CommandsPlugin from '@/extensions/slash-command/slash-command-plugin';
+import CommandsPlugin from '@/tiptap-extensions/slash-command/slash-command-plugin';
 import { Markdown } from 'tiptap-markdown';
-import { AIassistantNode } from '@/extensions/ai-generate/ai-generate-node';
+import { AIassistantNode } from '@/tiptap-extensions/ai-generate/ai-generate-node';
 import Underline from '@tiptap/extension-underline';
 import TextStyle from '@tiptap/extension-text-style';
 import FontFamily from '@tiptap/extension-font-family';
-import { FontSize } from '@/extensions/fontsize/font-size';
+import { FontSize } from '@/tiptap-extensions/fontsize/font-size';
 import Superscript from '@tiptap/extension-superscript';
 import Subscript from '@tiptap/extension-subscript';
 import TextAlign from '@tiptap/extension-text-align';
 import { Color } from '@tiptap/extension-color';
 import Placeholder from '@tiptap/extension-placeholder';
-import GlobalDragHandle from '@/extensions/drag-handle/drag-handle';
+import GlobalDragHandle from '@/tiptap-extensions/drag-handle/drag-handle';
 import Dropcursor from '@tiptap/extension-dropcursor';
-import { AiSuggestion } from '@/extensions/ai-suggestion/ai-suggestion';
-import { rules } from '@/extensions/ai-suggestion/rules';
+import { AiSuggestion } from '@/tiptap-extensions/ai-suggestion/ai-suggestion';
+import { rules } from '@/tiptap-extensions/ai-suggestion/rules';
 
 export const useTemplateEditor: (
   content: string,
@@ -38,6 +38,7 @@ export const useTemplateEditor: (
         rules: rules,
         loadOnStart: false,
         reloadOnUpdate: true,
+        debounceTimeout: 5000,
       }),
       Placeholder.configure({
         placeholder: 'type / for commands ',
