+++ src/tiptap-extensions/ai-generate/ai-generate-component.tsx
