+++ src/tiptap-extensions/ai-generate/ai-generate-node.ts
