+++ src/tiptap-extensions/ai-suggestion/rules.ts
