+++ src/tiptap-extensions/ai-suggestion/suggestion/rules-card-component.tsx
@@ -1,13 +1,4 @@
-import { useEditorStore } from '@/app/store/use-editor-store';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
-import { useState, useEffect } from 'react';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 
 interface Rule {
   id: string;
