+++ src/tiptap-extensions/ai-suggestion/suggestion/suggestion-panel-item.tsx
