+++ src/tiptap-extensions/ai-suggestion/suggestion/suggestion-panel.tsx
