+++ src/tiptap-extensions/document/document.ts
@@ -1,0 +1,7 @@
+import { Document as TiptapDocument } from '@tiptap/extension-document';
+
+export const Document = TiptapDocument.extend({
+  content: 'block+',
+});
+
+export default Document;
