+++ src/tiptap-extensions/drag-handle/clipboard-serializer.ts
