+++ src/tiptap-extensions/drag-handle/drag-handle.ts
