+++ src/tiptap-extensions/draggable-node.tsx
