+++ src/tiptap-extensions/paragraph/paragraph.ts
