+++ src/tiptap-extensions/slash-command/slash-command-plugin.ts
