+++ src/tiptap-extensions/table/table.ts
@@ -1,6 +1,6 @@
-import TipTapParagraph from "@tiptap/extension-paragraph";
-import { ReactNodeViewRenderer } from "@tiptap/react";
-import { DraggableNode } from "../draggable-node";
+import TipTapParagraph from '@tiptap/extension-paragraph';
+import { ReactNodeViewRenderer } from '@tiptap/react';
+import { DraggableNode } from '../draggable-node';
 
 export const Paragraph = TipTapParagraph.extend({
   draggable: true,
